requirements.txt
psutil>=5.8.0
Pillow>=9.0.0
python-pam>=2.0.0
